Make signing time methods static and extract to SigningInformationUtil
